requirements.txt
blessed==1.38.0
browser-cookie3==0.20.1
certifi==2026.2.25
charset-normalizer==3.4.7
cloudscraper==1.2.71
commonmark==0.9.1
idna==3.11
jeepney==0.9.0
lxml==5.4.0
lz4==4.4.5
measurement==3.2.2
mpmath==1.3.0
myfitnesspal==2.1.2
psycopg2-binary==2.9.11
pycryptodomex==3.23.0
pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
requests==2.33.1
requests-toolbelt==1.0.0
rich==12.6.0
six==1.17.0
sympy==1.14.0
typing-extensions==4.12.2
urllib3==2.6.3
wcwidth==0.6.0
